+++ NEWS.md
@@ -1,3 +1,4 @@
 # master
 
 * Store unit name
+* Fixed issue with binary prefix parsing in regards to JEDEC